Refactor: Sticker model separate
- 자주 사용되는 Sticker 모델 분리
