+++ app/src/main/java/com/sgs/devcamp2/flametalk_android/network/request/sign/ProfileUpdateRequest.kt
@@ -2,6 +2,7 @@ package com.sgs.devcamp2.flametalk_android.network.request.sign
 
 import androidx.annotation.Keep
 import com.google.gson.annotations.SerializedName
+import com.sgs.devcamp2.flametalk_android.data.model.Sticker
 
 /**
  * @author 박소연
@@ -23,7 +24,7 @@ import com.google.gson.annotations.SerializedName
  * stickerId	Integer	스티커 id
  * positionX	Double	스티커 x 좌표
  * positionY	Double	스티커 y 좌표
-*/
+ */
 
 @Keep
 data class ProfileUpdateRequest(
@@ -39,14 +40,4 @@ data class ProfileUpdateRequest(
     val description: String?,
     @SerializedName("isDefault")
     val isDefault: Boolean
-) {
-    @Keep
-    data class Sticker(
-        @SerializedName("stickerId")
-        val stickerId: Int,
-        @SerializedName("positionX")
-        val positionX: Double,
-        @SerializedName("positionY")
-        val positionY: Double
-    )
-}
+)
